requirements.txt
fastapi
uvicorn[standard]
boto3
python-dotenv
jinja2
aiofiles
